added user create func

+++ Server/migrations/2021-02-17-044649_create_mention/up.sql
@@ -1,8 +1,8 @@
 -- Your SQL goes here
 CREATE TABLE `mentions`
 (
-    `entry`  numeric NOT NULL,
-    `person` numeric NOT NULL,
+    `entry`  integer NOT NULL,
+    `person` integer NOT NULL,
 
     PRIMARY KEY (`entry`, `person`),
     FOREIGN KEY  (`entry`) REFERENCES `entries` (`id`),
