+++ Server/migrations/2021-02-17-044755_create_simp/up.sql
@@ -1,11 +1,11 @@
 -- Your SQL goes here
 CREATE TABLE `simping`
 (
-    `person`   numeric NOT NULL,
-    `entry`    numeric NOT NULL,
-    `amount`   numeric NOT NULL,
+    `person`   integer NOT NULL,
+    `entry`    integer NOT NULL,
+    `amount`   integer NOT NULL,
     `currency` text    NOT NULL,
-    `is_worth` numeric NOT NULL,
+    `is_worth` integer NOT NULL,
 
     PRIMARY KEY (`person`, `entry`),
     FOREIGN KEY  (`person`) REFERENCES `person` (`id`),
